Move directory client to /v2
