+++ client/directory/v2/config_test.go
@@ -1,0 +1,139 @@
+package directory // nolint:testpackage
+
+import (
+	"context"
+	"encoding/json"
+	"testing"
+
+	asserts "github.com/stretchr/testify/assert"
+
+	"github.com/aserto-dev/go-aserto/client"
+	"github.com/aserto-dev/go-aserto/client/directory/internal"
+)
+
+const (
+	base      = `{"address": "localhost:8282"}`
+	noBase    = `{"reader": {"address": "localhost:9292"}, "writer": {"address": "localhost:9393"}}`
+	overrides = `{"address": "localhost:8282", "reader": {"address": "localhost:9292"}}`
+)
+
+func TestUnmarshalConfig(t *testing.T) {
+	t.Run("base only", func(t *testing.T) {
+		assert := asserts.New(t)
+
+		cfg := Config{}
+		if err := json.Unmarshal([]byte(base), &cfg); err != nil {
+			assert.FailNow("failed to unmarshal config", err)
+		}
+
+		assert.NotNil(cfg.Config)
+		assert.Equal("localhost:8282", cfg.Address)
+		assert.Nil(cfg.Reader)
+
+		assert.NoError(cfg.Validate())
+	})
+
+	t.Run("no base", func(t *testing.T) {
+		assert := asserts.New(t)
+
+		var cfg Config
+		if err := json.Unmarshal([]byte(noBase), &cfg); err != nil {
+			assert.FailNow("failed to unmarshal config", err)
+		}
+
+		assert.Nil(cfg.Config)
+		assert.Nil(cfg.Importer)
+		assert.Nil(cfg.Exporter)
+		assert.NotNil(cfg.Reader)
+		assert.NotNil(cfg.Writer)
+		assert.Equal("localhost:9292", cfg.Reader.Address)
+		assert.Equal("localhost:9393", cfg.Writer.Address)
+	})
+
+	t.Run("overrides", func(t *testing.T) {
+		assert := asserts.New(t)
+
+		var cfg Config
+		if err := json.Unmarshal([]byte(overrides), &cfg); err != nil {
+			assert.FailNow("failed to unmarshal config", err)
+		}
+
+		assert.NotNil(cfg.Config)
+		assert.NotNil(cfg.Reader)
+		assert.Nil(cfg.Writer)
+		assert.Nil(cfg.Importer)
+		assert.Nil(cfg.Exporter)
+		assert.Equal("localhost:8282", cfg.Address)
+		assert.Equal("localhost:9292", cfg.Reader.Address)
+	})
+}
+
+func TestConnect(t *testing.T) {
+	ctx := context.Background()
+
+	t.Run("base only", func(t *testing.T) { // nolint:dupl
+		assert := asserts.New(t)
+
+		conns, counter := mockConns()
+
+		cfg := &Config{
+			Config: &client.Config{Address: "localhost:8282"},
+		}
+
+		dir, err := connect(ctx, conns, cfg)
+		assert.NoError(err)
+		assert.NotNil(dir)
+		assert.NotNil(dir.Reader)
+		assert.NotNil(dir.Writer)
+		assert.NotNil(dir.Importer)
+		assert.NotNil(dir.Exporter)
+		assert.Equal(1, counter.Count)
+	})
+
+	t.Run("base with overrides", func(t *testing.T) {
+		assert := asserts.New(t)
+
+		conns, counter := mockConns()
+
+		cfg := &Config{
+			Config: &client.Config{Address: "localhost:1234"},
+			Reader: &client.Config{Address: "localhost:4321"},
+		}
+
+		dir, err := connect(ctx, conns, cfg)
+		assert.NoError(err)
+		assert.NotNil(dir)
+		assert.NotNil(dir.Reader)
+		assert.NotNil(dir.Writer)
+		assert.NotNil(dir.Importer)
+		assert.NotNil(dir.Exporter)
+		assert.Equal(2, counter.Count)
+	})
+
+	t.Run("no base", func(t *testing.T) { // nolint:dupl
+		assert := asserts.New(t)
+
+		conns, counter := mockConns()
+
+		cfg := &Config{
+			Reader: &client.Config{Address: "localhost:9292"},
+		}
+
+		dir, err := connect(ctx, conns, cfg)
+		assert.NoError(err)
+		assert.NotNil(dir)
+		assert.NotNil(dir.Reader)
+		assert.Nil(dir.Writer)
+		assert.Nil(dir.Importer)
+		assert.Nil(dir.Exporter)
+		assert.Equal(1, counter.Count)
+	})
+}
+
+func mockConns() (*internal.Connections, *internal.ConnectCounter) {
+	counter := &internal.ConnectCounter{}
+	conns := internal.NewConnections()
+	conns.Connect = counter.Connect
+
+	return conns, counter
+}
